+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# (2023-04-10)
+# (2023-06-20)
 
 ### Bug Fixes
 
@@ -12,6 +12,7 @@
 ### Features
 
 - add longestCommonSubstring regularExpressionMatching ([c1ec6ca](https://github.com/air-supply94/js-algorithm/commit/c1ec6ca9180adf00dc15629676b758911d2545e8))
+- class 增加 implements ([f045dab](https://github.com/air-supply94/js-algorithm/commit/f045dabbb0c9761c1d722512f22944f7fff28647))
 - delete stylelint ([b84d971](https://github.com/air-supply94/js-algorithm/commit/b84d97118b65bc9769a739917b4718444e287e9f))
 - remove DoubleLinkedList reverse extra code ([4130002](https://github.com/air-supply94/js-algorithm/commit/413000283cc5024e481819919811bc4806cd5bf8))
 - rename file ([b03f0d1](https://github.com/air-supply94/js-algorithm/commit/b03f0d15ca04ebbfeb0f6a54aa3e1a0c4785e768))
@@ -232,7 +233,7 @@
 - **dp:** update regularExpressionMatching ([4651010](https://github.com/air-supply94/js-algorithm/commit/4651010ab2b61c7657f7191f58fff1e644348604))
 - **dp:** update superEggDrop ([7c9419c](https://github.com/air-supply94/js-algorithm/commit/7c9419cc1cab44bae2336ebaa75246b260d768ec))
 - **dp:** 优化 changeCoins 空间复杂度 ([30c7cf6](https://github.com/air-supply94/js-algorithm/commit/30c7cf66b299044cf46de2833235a275e2ee2fa7))
-- **dp:** 更新实现 ([a6b8046](https://github.com/air-supply94/js-algorithm/commit/a6b8046b6daa59e04a13a3dca711da847c509223))
+- **dp:** 更新实现 ([d8b5dda](https://github.com/air-supply94/js-algorithm/commit/d8b5ddae8d0e834eaaa291251f3ad58ccb864f37))
 - **dp:** 更新实现 ([65db4a2](https://github.com/air-supply94/js-algorithm/commit/65db4a266e485aab1b4ef1aa477d09b9ccf26191))
 - **dp:** 更新实现 ([e019f0f](https://github.com/air-supply94/js-algorithm/commit/e019f0f8ad4209e90a7a6918415ad7cb048dbcf3))
 - **dp:** 更新实现 ([427dd1b](https://github.com/air-supply94/js-algorithm/commit/427dd1b932c525d88cfd8f21d963f1190950dded))